requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
playwright==1.48.0
python-dotenv==1.0.0
rich==13.7.0
flask==3.0.0
anthropic==0.39.0
